feat(core): add image encoder

+++ src/vec3.rs
@@ -1,17 +1,17 @@
 /// Vec3 is a 3D vector.
 #[derive(Clone, Copy)]
 pub struct Vec3 {
-    comp: [f64; 3],
+    comp: [f32; 3],
 }
 
 impl Vec3 {
     /// Constructor.
-    fn new(x: f64, y: f64, z: f64) -> Self {
+    fn new(x: f32, y: f32, z: f32) -> Self {
         Vec3 { comp: [x, y, z] }
     }
 
     /// mag returns the magnitude/length of the vector.
-    fn mag(&self) -> f64 {
+    fn mag(&self) -> f32 {
         (self.x() * self.x() + self.y() * self.y() + self.z() * self.z()).sqrt()
     }
 
@@ -21,7 +21,7 @@ impl Vec3 {
     }
 
     // Dot product with another vector
-    fn dot(&self, other: &Vec3) -> f64 {
+    fn dot(&self, other: &Vec3) -> f32 {
         self.x() * other.x() + self.y() * other.y() + self.z() * other.z()
     }
 
@@ -36,17 +36,17 @@ impl Vec3 {
 
     // Helper methods for component access
     #[inline]
-    fn x(&self) -> f64 {
+    fn x(&self) -> f32 {
         self.comp[0]
     }
 
     #[inline]
-    fn y(&self) -> f64 {
+    fn y(&self) -> f32 {
         self.comp[1]
     }
 
     #[inline]
-    fn z(&self) -> f64 {
+    fn z(&self) -> f32 {
         self.comp[2]
     }
 }
@@ -79,19 +79,19 @@ impl std::ops::Sub<Vec3> for Vec3 {
 }
 
 /// Operator overload for mutliplication.
-impl std::ops::Mul<f64> for Vec3 {
+impl std::ops::Mul<f32> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, arg: f64) -> Vec3 {
+    fn mul(self, arg: f32) -> Vec3 {
         Vec3::new(self.x() * arg, self.y() * arg, self.z() * arg)
     }
 }
 
 /// Operator overload for division.
-impl std::ops::Div<f64> for Vec3 {
+impl std::ops::Div<f32> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, arg: f64) -> Vec3 {
+    fn div(self, arg: f32) -> Vec3 {
         self * (1.0 / arg)
     }
 }
